# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Steel Fabrication MRP API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ncrs:
    get:
      operationId: listNcrs
      tags: [qc]
      summary: List nonconformance reports
      parameters:
        - name: jobId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, closed]
      responses:
        "200":
          description: NCRs, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ncr"
    post:
      operationId: createNcr
      tags: [qc]
      summary: Create a nonconformance report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NcrInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ncr"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /ncrs/{ncrId}:
    patch:
      operationId: updateNcr
      tags: [qc]
      summary: Update an NCR (disposition, root cause, approval, closure)
      description: >
        Closing an NCR requires a disposition. Approval stamps the caller
        as approver.
      parameters:
        - name: ncrId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NcrUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ncr"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /substitution-requests:
    get:
      operationId: listSubstitutionRequests
      tags: [qc]
      summary: List material substitution requests
      parameters:
        - name: jobId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Substitution requests, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubstitutionRequest"
    post:
      operationId: createSubstitutionRequest
      tags: [qc]
      summary: Create a substitution request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubstitutionRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubstitutionRequest"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /substitution-requests/{requestId}:
    patch:
      operationId: updateSubstitutionRequest
      tags: [qc]
      summary: Update / approve / reject a substitution request
      description: >
        Approval is blocked for customer-specified or safety-critical
        substitutions until customer concurrence is recorded.
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubstitutionRequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubstitutionRequest"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Gate violation (customer concurrence required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
...
  schemas:
    Ncr:
      type: object
      required:
        - id
        - number
        - source
        - description
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        number:
          type: string
        source:
          type: string
          enum: [receiving, in_process, final, post_delivery]
        description:
          type: string
        jobId:
          type: ["integer", "null"]
        jobNumber:
          type: ["string", "null"]
        assemblyId:
          type: ["integer", "null"]
        assemblyMark:
          type: ["string", "null"]
        purchaseOrderId:
          type: ["integer", "null"]
        poNumber:
          type: ["string", "null"]
        disposition:
          type: ["string", "null"]
          enum: [rework, scrap, accept_with_deviation, null]
        dispositionNotes:
          type: ["string", "null"]
        rootCause:
          type: ["string", "null"]
        approvedBy:
          type: ["integer", "null"]
        approvedByName:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, closed]
        closedAt:
          type: ["string", "null"]
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    NcrInput:
      type: object
      required: [source, description]
      properties:
        source:
          type: string
          enum: [receiving, in_process, final, post_delivery]
        description:
          type: string
          minLength: 1
        jobId:
          type: ["integer", "null"]
        assemblyId:
          type: ["integer", "null"]
        purchaseOrderId:
          type: ["integer", "null"]
    NcrUpdate:
      type: object
      properties:
        description:
          type: string
        disposition:
          type: ["string", "null"]
          enum: [rework, scrap, accept_with_deviation, null]
        dispositionNotes:
          type: ["string", "null"]
        rootCause:
          type: ["string", "null"]
        approve:
          type: boolean
          description: Stamp the caller as approver
        close:
          type: boolean
          description: Close the NCR (requires a disposition)
        reopen:
          type: boolean
    SubstitutionRequest:
      type: object
      required:
        - id
        - number
        - originalSpec
        - proposedSubstitution
        - type
        - engineeringRationale
        - customerSpecified
        - safetyCritical
        - customerConcurrence
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        number:
          type: string
        jobId:
          type: ["integer", "null"]
        jobNumber:
          type: ["string", "null"]
        assemblyId:
          type: ["integer", "null"]
        assemblyMark:
          type: ["string", "null"]
        originalSpec:
          type: string
        proposedSubstitution:
          type: string
        type:
          type: string
          enum: [like_for_like, equivalent, upgrade, downgrade]
        engineeringRationale:
          type: string
        customerSpecified:
          type: boolean
        safetyCritical:
          type: boolean
        customerConcurrence:
          type: boolean
        concurrenceReference:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        approvedBy:
          type: ["integer", "null"]
        approvedByName:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        executionReference:
          type: ["string", "null"]
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    SubstitutionRequestInput:
      type: object
      required: [originalSpec, proposedSubstitution, type, engineeringRationale]
      properties:
        originalSpec:
          type: string
          minLength: 1
        proposedSubstitution:
          type: string
          minLength: 1
        type:
          type: string
          enum: [like_for_like, equivalent, upgrade, downgrade]
        engineeringRationale:
          type: string
          minLength: 1
        jobId:
          type: ["integer", "null"]
        assemblyId:
          type: ["integer", "null"]
        customerSpecified:
          type: boolean
        safetyCritical:
          type: boolean
    SubstitutionRequestUpdate:
      type: object
      properties:
        customerConcurrence:
          type: boolean
        concurrenceReference:
          type: ["string", "null"]
        executionReference:
          type: ["string", "null"]
        approve:
          type: boolean
        reject:
          type: boolean
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...